pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seedance"
version = "0.1.0"
description = "Volces API Client Library"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0",
    "urllib3>=1.26.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["seedance*"]